feat: 텍스트 속성을 spread 하게 하여 텍스트 노드의 자식으로 블록 노드를 두지 못하도록 수정
