+++ package.json
@@ -37,6 +37,7 @@
     "fast-xml-parser": "^4.0.12",
     "framer-motion": "^7.10.2",
     "hotkeys-js": "^3.10.1",
+    "html-tags": "^3.2.0",
     "immer": "^9.0.16",
     "is-hotkey": "^0.2.0",
     "is-url": "^1.2.4",
