+++ src/modules/import/wordpress/parse.test.ts
@@ -2,11 +2,19 @@ import fs from "fs";
 import { parse } from "parse5";
 
 import { getPublishedPost, htmlToSlateFragment, parseXml } from "./parse";
-import sanitizeCode from "./sanitize-code";
+import sanitize from "./sanitize";
 
 const xmlFile = fs.readFileSync(
   "./src/modules/import/wordpress/__mocks__/underthepencil.WordPress.2022-12-17.xml"
 );
+
+vi.mock("nanoid", () => {
+  const nanoid = vi.fn(() => {
+    return `mocked-nanoid-random-id`;
+  });
+
+  return { nanoid };
+});
 
 describe("parseXml", () => {
   test("제대로 동작해야 함", async () => {
@@ -51,12 +59,110 @@ describe("htmlToSlateFragment", () => {
     expect(result).toMatchSnapshot();
   });
 
+  test("Text 끼리 중첩되었을 때 잘 동작해야 함", () => {
+    const html = parse(`
+    <p>123<em>h<strong>e<s>l</s>l</strong>o</em>h1</p>
+    `);
+    const result = htmlToSlateFragment(html);
+    expect(result).toMatchInlineSnapshot(`
+      [
+        {
+          "children": [
+            {
+              "text": "123",
+              "type": "TEXT",
+            },
+            {
+              "italic": true,
+              "text": "h",
+              "type": "TEXT",
+            },
+            {
+              "bold": true,
+              "italic": true,
+              "text": "e",
+              "type": "TEXT",
+            },
+            {
+              "bold": true,
+              "italic": true,
+              "strikethrough": true,
+              "text": "l",
+              "type": "TEXT",
+            },
+            {
+              "bold": true,
+              "italic": true,
+              "text": "l",
+              "type": "TEXT",
+            },
+            {
+              "italic": true,
+              "text": "o",
+              "type": "TEXT",
+            },
+            {
+              "text": "h1",
+              "type": "TEXT",
+            },
+          ],
+          "type": "PARAGRAPH",
+        },
+      ]
+    `);
+  });
+
+  test("텍스트 노드 안에 링크 있을 떄 처리", () => {
+    const html = parse(`
+    <p>123h<strong>easdf <a href="http://naver.com">adf</a> asdf</strong>asdf asdf</p>
+    `);
+    const result = htmlToSlateFragment(html);
+    expect(result).toMatchInlineSnapshot(`
+      [
+        {
+          "children": [
+            {
+              "text": "123h",
+              "type": "TEXT",
+            },
+            {
+              "bold": true,
+              "text": "easdf ",
+              "type": "TEXT",
+            },
+            {
+              "children": [
+                {
+                  "bold": true,
+                  "text": "adf",
+                  "type": "TEXT",
+                },
+              ],
+              "type": "LINK",
+              "url": "http://naver.com",
+            },
+            {
+              "bold": true,
+              "text": " asdf",
+              "type": "TEXT",
+            },
+            {
+              "text": "asdf asdf",
+              "type": "TEXT",
+            },
+          ],
+          "type": "PARAGRAPH",
+        },
+      ]
+    `);
+  });
+
   describe("Wordpress Content", async () => {
     const wpJson = await parseXml(xmlFile);
     const wpItems = getPublishedPost(wpJson);
     wpItems.forEach((post) => {
       it(`실제 워드프레스 포스팅을 잘 변환해야 함: ${post.post_id}`, () => {
-        const htmlParsed = parse(sanitizeCode(post.content));
+        const htmlParsed = parse(sanitize(post.content));
         const result = htmlToSlateFragment(htmlParsed);
         expect(result).toMatchSnapshot();
       });
